Initial commit
From now on every change MUST be tracked via git.
For every new feature, you have to create a new branch.
Every PR must be linked to an issue that describes the change.

Issue and PR are different! The Issue describes it before coding.
The PR after codeing. Therefore they have not the same content.

The Issue describes what should be changed and why.
The PR describes how the change was done.